requirements.txt
numpy
Pillow
torch>=1.2
tqdm
torchvision